backend/requirements.txt
# GreenGate Backend - Dependencies
# Python 3.11+

# NumPy (MUST be < 2.0 for Shapely/GeoAlchemy2 compatibility)
numpy<2.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9  # Para Alembic (sync)
geoalchemy2==0.14.3
alembic==1.13.1

# Geospatial (apenas o essencial - evita "maldição do GDAL")
shapely==2.0.2
pyproj==3.6.1
# geopandas e fiona removidos - não são essenciais e causam problemas de deps

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client (para APIs externas)
httpx==0.26.0
aiohttp==3.9.1

# Cache & Rate Limiting
redis==5.0.1

# PDF Generation
reportlab==4.0.8
# weasyprint removido - deps complexas, reportlab é suficiente

# Utils
python-dotenv==1.0.0
orjson==3.9.10
tenacity==8.2.3
psutil==5.9.8

# Logging estruturado
structlog==24.1.0

# Dev & Testing (não vão para produção se usar multi-stage build)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
qrcode[pil]==7.4.2
